requirements.txt
pandas==2.2.3
python-binance==1.0.27
python-dotenv==1.0.1